sender/requirements.txt
azure-servicebus==7.8.2
termcolor==2.2.0